refactor(festivals): move festival activity admin UI into shared components
Relocate activity tables, proof modals, participant selection, export button,
and related pieces from the dashboard route folder into
app/components/festivals/festival_activities. Rename activity-details to
admin-activity-details and move the coupon book preview into
coupon-book-preview-client for the couponbook review route.
