+++ app/components/festivals/festival_activities/participant-card.tsx
