+++ app/components/festivals/festival_activities/text-proof-modal.tsx
